+++ R/dirichletprocess.R
@@ -3,7 +3,7 @@
 #' @description Create, fit and take posterior samples from a Dirichlet process.
 #'
 #'
-#' @importFrom stats dbeta dbinom dgamma dnorm dt dunif dweibull rWishart rbeta rgamma rnorm runif var quantile
+#' @importFrom stats dbeta dbinom dgamma dnorm dt dunif dweibull dexp rWishart rbeta rgamma rnorm runif var quantile optim
 #'
 #' @importFrom utils setTxtProgressBar txtProgressBar
 #'