+++ R/mixing_distribution_penalised_likelihood.R
@@ -1,0 +1,8 @@
+
+PenalisedLikelihood <- function(mdObj, x){
+  UseMethod("PenalisedLikelihood", mdObj)
+}
+
+PenalisedLikelihood.default <- function(mdObj, x){
+  return(PriorDraw(mdObj, 1))
+}