+++ R/mixing_distribution_posterior_draw.R
@@ -11,7 +11,7 @@ PosteriorDraw <- function(mdObj, x, n = 1, ...) UseMethod("PosteriorDraw", mdObj
 PosteriorDraw.nonconjugate <- function(mdObj, x, n = 1, start_pos, ...) {
 
   if (missing(start_pos)) {
-    start_pos <- PriorDraw(mdObj)
+    start_pos <- PenalisedLikelihood(mdObj, x)
   }
 
   mh_result <- MetropolisHastings(mdObj, x, start_pos, no_draws = n)
@@ -19,8 +19,8 @@ PosteriorDraw.nonconjugate <- function(mdObj, x, n = 1, start_pos, ...) {
   theta <- vector("list", length(mh_result))
 
   for (i in seq_along(mh_result$parameter_samples)) {
-    theta[[i]] <- array(mh_result$parameter_samples[[i]], dim = c(dim(mh_result$parameter_sample[[i]])[1:2],
-      n))
+    theta[[i]] <- array(mh_result$parameter_samples[[i]],
+                        dim = c(dim(mh_result$parameter_sample[[i]])[1:2], n))
   }
 
   return(theta)